Add C.J instruction
C.JAL and C.J were meant to be in this commit as complementary instructions,
but C.JAL is not supported on RV64. TYPE_MID was intended to use C.JAL to
reduce its size by 2 bytes, but due to RV64 limitations, only C.J is
included in this commit.